updated .gitignore and refactored as per naming convention
