+++ .gitignore
@@ -1,5 +1,5 @@
-/.classpath
+.classpath
 .project
-/.settings/org.eclipse.jdt.core.prefs
-/.settings/org.eclipse.m2e.core.prefs
-/target
+target/
+.settings/*
+test-output/*
