+++ src/main/java/api/BaseSpec.java
@@ -1,4 +1,4 @@
-package adApiTest;
+package api;
 
 import java.util.Map;
